architecture: refactored app archi
