+++ signit-interface/src/styles/traduction.css
